requirements.txt
asgiref==3.9.1
Django==5.2.5
django-extensions==4.1
django-tenant-users==2.1.1
django-tenants==3.8.0
psycopg2==2.9.10
python-decouple==3.8
sqlparse==0.5.3
typing_extensions==4.14.1
